// resources/js/modules/user/redux/user.types.ts
enum UserTypes {
    USER_ACTION_START = "USER_ACTION_START",
    USER_GET_SUCCESS = "USER_GET_SUCCESS",
    USER_GET_FAILED = "USER_GET_FAILED",
    USER_CREATE_SUCCESS = "USER_CREATE_SUCCESS",
    USER_CREATE_FAILED = "USER_CREATE_FAILED",
    USER_GET_BY_ID_SUCCESS = "USER_GET_BY_ID_SUCCESS",
    USER_GET_BY_ID_FAILED = "USER_GET_BY_ID_FAILED",
    USER_UPDATE_SUCCESS = "USER_UPDATE_SUCCESS",
    USER_UPDATE_FAILED = "USER_UPDATE_FAILED",
    USER_DELETE_SUCCESS = "USER_DELETE_SUCCESS",
    USER_DELETE_FAILED = "USER_DELETE_FAILED",
    USER_ACTION_END = "USER_ACTION_END",
}
export default UserTypes;